+++ frontend/mulighetsrommet-api-client/package.json
@@ -6,7 +6,7 @@
   "types": "build/index.ts",
   "scripts": {
     "postinstall": "npm run build",
-    "build": "openapi --useOptions --input ../../mulighetsrommet-api/src/main/resources/web/openapi.yml --output ./build"
+    "build": "openapi --useOptions --name MulighetsrommetClient --input ../../mulighetsrommet-api/src/main/resources/web/openapi.yml --output ./build"
   },
   "devDependencies": {
     "openapi-typescript-codegen": "^0.23.0"
