+++ frontend/mulighetsrommet-veileder-flate/src/core/api/clients.ts
@@ -1,0 +1,3 @@
+import { MulighetsrommetClient } from 'mulighetsrommet-api-client';
+
+export const mulighetsrommetClient = new MulighetsrommetClient();
